baseline/requirements.txt
accelerate==1.6.0
beautifulsoup4==4.13.4
faiss_gpu==1.7.2
huggingface_hub==0.30.1
jsonlines==4.0.0
nervaluate==0.2.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.4
openai==1.78.1
pandas==2.2.3
peft==0.15.2
rdflib==7.1.4
regex==2024.11.6
scikit_learn==1.6.1
sentence_transformers==4.1.0
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
Unidecode==1.4.0
vllm==0.8.5
